WIP | refactor. landing page of inventory module
