+++ index.scss
@@ -1,0 +1,1 @@
+@import '~carbon-components/scss/globals/scss/styles.scss';
