+++ package.json
@@ -1,12 +1,12 @@
 {
 	"name": "@bahmni/inventory-app",
-  "browser": "public/main.js",
+	"browser": "public/main.js",
 	"version": "1.0.0",
 	"description": "Bahmni Inventory Frontend",
 	"main": "index.js",
 	"scripts": {
-		"start": "webpack-dev-server .",
-		"build": "webpack .",
+		"start": "rimraf public/main.js && webpack-dev-server .",
+		"build": "rimraf public/main.js && webpack .",
 		"test": "jest --passWithNoTests",
 		"test:watch": "jest --watch",
 		"lint": "eslint src",
@@ -14,12 +14,13 @@
 	},
 	"author": "",
 	"license": "ISC",
-  "peerDependencies": {
-    "swr": "^1.1.2",
-    "@openmrs/esm-framework": "3.4.1-pre.189",
-    "react": "^17.0.2",
-    "carbon-components-react": "^7.0.0"
-  },
+	"peerDependencies": {
+		"@openmrs/esm-framework": "3.4.1-pre.189",
+		"carbon-components-react": "^7.0.0",
+		"carbon-icons": "^7.0.7",
+		"react": "^17.0.2",
+		"swr": "^1.1.2"
+	},
 	"devDependencies": {
 		"@babel/cli": "^7.21.5",
 		"@babel/core": "^7.22.1",
@@ -28,8 +29,10 @@
 		"@babel/preset-env": "^7.22.4",
 		"@babel/preset-react": "^7.22.3",
 		"@babel/runtime": "^7.22.3",
+		"@openmrs/esm-framework": "3.4.1-pre.189",
 		"babel-eslint": "^10.1.0",
 		"babel-loader": "^8.3.0",
+		"carbon-components-react": "^7.0.0",
 		"eslint": "^8.41.0",
 		"eslint-config-airbnb-base": "^15.0.0",
 		"eslint-config-prettier": "^8.8.0",
@@ -37,19 +40,23 @@
 		"eslint-plugin-jsx-a11y": "^6.7.1",
 		"eslint-plugin-react": "^7.32.2",
 		"eslint-plugin-react-hooks": "^4.6.0",
-		"path": "^0.12.7",
 		"jest": "^29.5.0",
+		"openmrs": "^4.5.0",
+		"path": "^0.12.7",
+		"swr": "^1.1.2",
 		"webpack": "^5.85.0",
 		"webpack-cli": "^4.10.0",
 		"webpack-dev-server": "^4.15.0",
-    "@openmrs/esm-framework": "3.4.1-pre.189",
-    "swr": "^1.1.2",
-    "carbon-components-react": "^7.0.0",
-    "openmrs": "3.4.1-pre.189"
+		"style-loader": "^3.3.1",
+		"sass": "^1.44.0",
+    "sass-loader": "^12.3.0",
+    "css-loader": "^5.2.4"
 	},
 	"dependencies": {
+		"@carbon/icons-react": "^10.16.0",
+		"@carbon/react": "^1.1.0",
+		"carbon-components-react": "^7.0.0",
 		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
-    "carbon-components-react": "^7.0.0"
+		"react-dom": "^17.0.2"
 	}
 }
