+++ src/App.js
@@ -1,21 +1,24 @@
 import React from "react";
 import { InventoryLandingPage } from "./inventory/inventory-landing-page";
 import {Dropdown } from 'carbon-components-react'
+import 'carbon-components/scss/globals/scss/styles.scss';
 
-const App = () =>(
-    <div>
+const App = () =>{
+  const items = ['HWC1','HWC2','HWC3']
+  return  <div>
+      <h1>******Inventory*****</h1>
         {/* <Dropdown>HWC</Dropdown> */}
         <Dropdown
           id="hwc-list-dropdown"
           title="hwc list"
           items={items}
-          itemToString={data => data.person?.display}
+          itemToString={data => data}
           label="Select HWC"
         //   onChange={({selectedItem}) => setDoctor(selectedItem)}
-          selectedItem={"doctor"}
+          // selectedItem={"doctor"}
         />
-        <InventoryLandingPage></InventoryLandingPage>
+        {/* <InventoryLandingPage/> */}
     </div>
-    )
+}
 
 export default App
